Docker: run bundler before importing our code
Docker creates a intermediate image for every step in the Dockerfile.
Those images are reused whenever the context of the command did not
change (the copied files in this case). If we install our dependencies
first, and then import the rest of our project, Docker will be able to
reuse the intermediate image containing our already installed
dependencies.